# Conflicts:
#	TestProjects/HDRP_Tests/Assets/GraphicTests/Scenes/9x_Other/9700_CustomPass_FullScreen.unity
#	TestProjects/HDRP_Tests/Assets/GraphicTests/Scenes/9x_Other/9700_CustomPass_FullScreen/Unlit.mat
#	TestProjects/HDRP_Tests/Assets/GraphicTests/Scenes/9x_Other/9701_CustomPass_DrawRenderers.unity
#	TestProjects/HDRP_Tests/Assets/GraphicTests/Scenes/9x_Other/9702_CustomPass_API.unity
#	TestProjects/HDRP_Tests/Assets/GraphicTests/Scenes/9x_Other/9702_CustomPass_API/9702_LitUV.mat
#	TestProjects/HDRP_Tests/Assets/GraphicTests/Scenes/9x_Other/9702_CustomPass_API/9702_UnlitUVChecker.mat
#	TestProjects/HDRP_Tests/Assets/ReferenceImages/Linear/OSXEditor/Metal/None/9702_CustomPass_API.png
#	TestProjects/HDRP_Tests/Assets/ReferenceImages/Linear/OSXEditor/Metal/None/9702_CustomPass_API.png.meta
#	TestProjects/HDRP_Tests/Assets/ReferenceImages/Linear/WindowsEditor/Direct3D11/None/9702_CustomPass_API.png
#	TestProjects/HDRP_Tests/Assets/ReferenceImages/Linear/WindowsEditor/Direct3D11/None/9702_CustomPass_API.png.meta
#	TestProjects/HDRP_Tests/Assets/ReferenceImages/Linear/WindowsEditor/Vulkan/None/9702_CustomPass_API.png
#	TestProjects/HDRP_Tests/Assets/ReferenceImages/Linear/WindowsEditor/Vulkan/None/9702_CustomPass_API.png.meta
#	com.unity.render-pipelines.high-definition/CHANGELOG.md
#	com.unity.render-pipelines.high-definition/Runtime/Compositor/CustomClear.cs
#	com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/Camera/HDCamera.cs
#	com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDRenderPipeline.cs
#	com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDStringConstants.cs
#	com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/RenderPass/CustomPass/CustomPassContext.cs
#	com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/RenderPass/CustomPass/CustomPassRenderersUtils.shader
#	com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/RenderPass/CustomPass/CustomPassUtils.cs
#	com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/RenderPass/CustomPass/CustomPassUtils.shader
#	com.unity.render-pipelines.high-definition/Runtime/RenderPipelineResources/HDRenderPipelineResources.asset

+++ com.unity.render-pipelines.universal/Shaders/2D/Light2D-Shape.shader
@@ -40,23 +40,23 @@ Shader "Hidden/Light2D-Shape"
             struct Varyings
             {
                 float4  positionCS  : SV_POSITION;
-                float4  color       : COLOR;
-                float2  uv          : TEXCOORD0;
+                half4   color       : COLOR;
+                half2   uv          : TEXCOORD0;
 
                 SHADOW_COORDS(TEXCOORD1)
                 NORMALS_LIGHTING_COORDS(TEXCOORD2, TEXCOORD3)
             };
 
-            float  _InverseHDREmulationScale;
-            float4 _LightColor;
-            float  _FalloffDistance;
-            float4 _FalloffOffset;
+            half    _InverseHDREmulationScale;
+            half4   _LightColor;
+            half    _FalloffDistance;
+            half4   _FalloffOffset;
 
 #ifdef SPRITE_LIGHT
             TEXTURE2D(_CookieTex);			// This can either be a sprite texture uv or a falloff texture
             SAMPLER(sampler_CookieTex);
 #else
-            float _FalloffIntensity;
+            half    _FalloffIntensity;
             TEXTURE2D(_FalloffLookup);
             SAMPLER(sampler_FalloffLookup);
 #endif
